COrrecio de errores

+++ src/main/java/com/miguelangel/taqueria/services/CategoriaService.java
@@ -7,14 +7,20 @@ import org.springframework.stereotype.Service;
 
 import com.miguelangel.taqueria.entities.Categoria;
 import com.miguelangel.taqueria.repositories.CategoriaRepository;
+import com.miguelangel.taqueria.repositories.TacoRepository;
 
 @Service
 public class CategoriaService {
 
     private final CategoriaRepository categoriaRepository;
+    private final TacoRepository tacoRepository;
 
-    public CategoriaService(CategoriaRepository categoriaRepository) {
+    public CategoriaService(
+            CategoriaRepository categoriaRepository,
+            TacoRepository tacoRepository) {
+
         this.categoriaRepository = categoriaRepository;
+        this.tacoRepository = tacoRepository;
     }
 
     public List<Categoria> listarTodas() {
@@ -29,7 +35,17 @@ public class CategoriaService {
         return categoriaRepository.save(categoria);
     }
 
-    public void eliminar(Integer id) {
+    public boolean eliminar(Integer id) {
+
+        if (!categoriaRepository.existsById(id)) {
+            return false;
+        }
+
+        if (tacoRepository.existsByCategoriaId(id)) {
+            return false;
+        }
+
         categoriaRepository.deleteById(id);
+        return true;
     }
 }
